requirements.txt
# MNIST Digit Classification Project - Requirements

# Core ML/DL Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model Persistence
joblib>=1.1.0

# Streamlit App Dependencies
streamlit>=1.25.0
streamlit-drawable-canvas>=0.9.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Jupyter Notebook
notebook>=6.4.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Optional but recommended
jupyter>=1.0.0
